+++ src/App/Shared/GoldEx.Shared/DTOs/Transactions/GetTransactionResponse.cs
@@ -4,13 +4,13 @@ using GoldEx.Shared.Enums;
 namespace GoldEx.Shared.DTOs.Transactions;
 
 public record GetTransactionResponse(Guid Id, 
-    int Number, 
+    long Number, 
     string Description,
     DateTime DateTime,
-    double? Credit,
+    decimal? Credit,
     UnitType? CreditUnit,
-    double? CreditRate,
-    double? Debit,
+    decimal? CreditRate,
+    decimal? Debit,
     UnitType? DebitUnit,
-    double? DebitRate,
+    decimal? DebitRate,
     GetCustomerResponse Customer);
